Revert: Use correct dependency name for Multiverse-Core
Reverts the previous commit.

The plugin name for Multiverse-Core is 'Multiverse-Core', not 'multiverse-core'. Using the lowercase name causes an `UnknownDependencyException`.

This commit changes the `depend` key in `plugin.yml` back to 'Multiverse-Core' to ensure the server's dependency loader can find the plugin.
