+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Tue Aug 12 15:43:41 UTC 2025
+gradle.version=8.5
