+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'endworldreset'
+rootProject.name = 'KartaWorldReset'
